# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Staff Application Portal API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusBot = {
  connected: boolean;
  tag?: string | null;
  uptime: number;
  guilds: number;
};

export interface HealthStatus {
  status: string;
  bot?: HealthStatusBot;
}

export interface AuthUser {
  id: string;
  username: string;
  avatar: string | null;
  discriminator: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export type ApplicationInputRole =
  (typeof ApplicationInputRole)[keyof typeof ApplicationInputRole];

export const ApplicationInputRole = {
  Moderator: "Moderator",
  Human_Resources: "Human Resources",
  Partnership: "Partnership",
} as const;

export interface ApplicationInput {
  /** @minLength 2 */
  discordUsername: string;
  /** @minLength 17 */
  discordId: string;
  role: ApplicationInputRole;
  /** @minimum 13 */
  age: number;
  timezone: string;
  /** @minLength 20 */
  experience: string;
  /** @minLength 20 */
  whyJoin: string;
  availability: string;
}

export interface Application {
  id: number;
  discordUsername: string;
  discordId: string;
  role: string;
  age: number;
  timezone: string;
  experience: string;
  whyJoin: string;
  availability: string;
  status: string;
  createdAt: string;
}

export interface ApplicationStats {
  total: number;
  pending: number;
  accepted: number;
  denied: number;
}

export interface ErrorResponse {
  error: string;
}