# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Staff Application Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  username: string;
  avatar: string | null;
  discriminator: string;
}